refactor handle_author

+++ Flask1/app.py
@@ -97,40 +97,33 @@ def handle_authors():
          except:
             abort(400, "UNIQUE constraint failed")   
 
-@app.get("/authors/<int:author_id>")
-def get_author(author_id):
+@app.route("/authors/<int:author_id>", methods=["GET", "PUT", "DELETE"])
+def handle_author(author_id):
       author = AuthorModel.query.get(author_id)
-      if author:
+      if not author:
+         abort(404, f"Author with id = {author_id} not found")
+      
+      if request.method == "GET":
          return jsonify(author.to_dict()), 200
-      abort(404, f"Author with id = {author_id} not found")
 
-@app.put("/authors/<int:author_id>")
-def edit_author(author_id):
-   new_data = request.json
-   author = AuthorModel.query.get(author_id)
-   if not author:
-      abort(404, f"Author with id = {author_id} not found")
-   
-    # Универсальный случай
-   for key, value in new_data.items():
-      setattr(author, key, value)
-   try:
-      db.session.commit()
-      return jsonify(author.to_dict()), 200
-   except:
-      abort(400, f"Database commit operation failed.")
+      if request.method == "PUT":
+         new_data = request.json
+         # Универсальный случай
+         for key, value in new_data.items():
+            setattr(author, key, value)
+         try:
+            db.session.commit()
+            return jsonify(author.to_dict()), 200
+         except:
+            abort(400, f"Database commit operation failed.") 
 
-@app.delete("/authors/<int:author_id>")
-def delete_author(author_id):
-   author = AuthorModel.query.get(author_id)
-   if not author:
-      abort(404, f"Author with id = {author_id} not found")
-   db.session.delete(author)
-   try:
-      db.session.commit()
-      return jsonify(message=f"Author with id = {r} deleted successfully"), 200
-   except:
-      abort(400, f"Database commit operation failed.")
+      if request.method == "DELETE":
+         db.session.delete(author)
+         try:
+            db.session.commit()
+            return jsonify(message=f"Author with id={author_id} deleted successfully"), 200
+         except:
+            abort(400, f"Database commit operation failed.")      
 
 @app.post("/authors/<int:author_id>/quotes")
 def create_quote(author_id):
